requirements.txt
torch
torchdiffeq
pyro-ppl 